refactor(gda): move scene and node groups into gda.commands (ADR-0040)
Each group module owns its models, renderers, descriptors, and Typer
bodies, exposing register(root); node imports SceneNode from scene
(sanctioned one-way edge). Registration order and the public surface
are unchanged; test imports follow the moves.
