+++ src/gda/cli.py
@@ -23,6 +23,7 @@ import typer
 from pydantic import ValidationError
 
 from gda import dispatch
+from gda.commands import node as node_commands, scene as scene_commands
 from gda.daemon_ops import (
     run_daemon_start_operation,
     run_daemon_status_operation,
@@ -113,24 +114,6 @@ from gda.models import (
     ProjectFindUnusedResourcesResult,
     ProjectStatisticsParams,
     ProjectStatisticsResult,
-    NodeAddParams,
-    NodeAddResult,
-    NodeConnectSignalParams,
-    NodeConnectSignalResult,
-    NodeDisconnectSignalParams,
-    NodeDisconnectSignalResult,
-    NodeDuplicateParams,
-    NodeDuplicateResult,
-    NodeGetParams,
-    NodeGetResult,
-    NodeListParams,
-    NodeListResult,
-    NodeMoveParams,
-    NodeMoveResult,
-    NodeRemoveParams,
-    NodeRemoveResult,
-    NodeSetParams,
-    NodeSetResult,
     ProjectAddAutoloadParams,
     ProjectAddAutoloadResult,
     ProjectAddInputActionParams,
@@ -162,16 +145,6 @@ from gda.models import (
     ScreenCaptureResult,
     ScreenFramesParams,
     ScreenFramesResult,
-    SceneCreateParams,
-    SceneCreateResult,
-    SceneDeleteParams,
-    SceneDeleteResult,
-    SceneGetExportsParams,
-    SceneGetExportsResult,
-    SceneGetParams,
-    SceneGetResult,
-    SceneListParams,
-    SceneListResult,
     ScriptAttachParams,
     ScriptAttachResult,
     ScriptCreateParams,
@@ -220,15 +193,6 @@ from gda.render import (
     render_input_mouse,
     render_input_sequence,
     render_logger_tail,
-    render_node_add,
-    render_node_connect_signal,
-    render_node_disconnect_signal,
-    render_node_duplicate,
-    render_node_list,
-    render_node_move,
-    render_node_properties,
-    render_node_remove,
-    render_node_set,
     render_perf_monitor,
     render_perf_monitors,
     render_project_add_autoload,
@@ -248,11 +212,6 @@ from gda.render import (
     render_resource_properties,
     render_resource_set,
     render_resource_uid,
-    render_scene_delete,
-    render_scene_exports,
-    render_scene_list,
-    render_scene_metadata,
-    render_scene_tree,
     render_screen_capture,
     render_screen_frames,
     render_script_attach,
@@ -285,16 +244,11 @@ app = typer.Typer(
     add_completion=False,
 )
 
-# The first domain command group (ADR-0005): commands acting on scene files.
-scene_app = typer.Typer(help="Act on Godot scene files (.tscn).", no_args_is_help=True)
-app.add_typer(scene_app, name="scene")
-
-# The node command group (issue #53): commands acting on nodes WITHIN a scene
-# file (load → locate → mutate → pack → save), so they stay headless.
-node_app = typer.Typer(
-    help="Act on nodes within a scene file (.tscn).", no_args_is_help=True
-)
-app.add_typer(node_app, name="node")
+# The scene and node groups own their sub-apps (ADR-0040) and mount them here,
+# at the same point in the sequence the old `add_typer` calls occupied, so the
+# registration order — and therefore `gda --help` — is unchanged.
+scene_commands.register(app)
+node_commands.register(app)
 
 # The script command group (issue #110): commands acting on .gd script files on
 # disk (write text / read text back), so they stay headless. C# (.cs) is out of
@@ -1558,106 +1512,6 @@ def screen_frames(
     )
 
 
-SCENE_CREATE_COMMAND: HeadlessCommand[SceneCreateResult] = HeadlessCommand(
-    operation="scene-create",
-    input_model=SceneCreateParams,
-    output_model=SceneCreateResult,
-    render=render_scene_metadata,
-)
-
-SCENE_GET_COMMAND: HeadlessCommand[SceneGetResult] = HeadlessCommand(
-    operation="scene-get",
-    input_model=SceneGetParams,
-    output_model=SceneGetResult,
-    render=render_scene_tree,
-)
-
-SCENE_GET_EXPORTS_COMMAND: HeadlessCommand[SceneGetExportsResult] = HeadlessCommand(
-    operation="scene-get-exports",
-    input_model=SceneGetExportsParams,
-    output_model=SceneGetExportsResult,
-    render=render_scene_exports,
-)
-
-SCENE_LIST_COMMAND: HeadlessCommand[SceneListResult] = HeadlessCommand(
-    operation="scene-list",
-    input_model=SceneListParams,
-    output_model=SceneListResult,
-    render=render_scene_list,
-)
-
-SCENE_DELETE_COMMAND: HeadlessCommand[SceneDeleteResult] = HeadlessCommand(
-    operation="scene-delete",
-    input_model=SceneDeleteParams,
-    output_model=SceneDeleteResult,
-    render=render_scene_delete,
-)
-
-NODE_ADD_COMMAND: HeadlessCommand[NodeAddResult] = HeadlessCommand(
-    operation="node-add",
-    input_model=NodeAddParams,
-    output_model=NodeAddResult,
-    render=render_node_add,
-)
-
-NODE_LIST_COMMAND: HeadlessCommand[NodeListResult] = HeadlessCommand(
-    operation="node-list",
-    input_model=NodeListParams,
-    output_model=NodeListResult,
-    render=render_node_list,
-)
-
-NODE_GET_COMMAND: HeadlessCommand[NodeGetResult] = HeadlessCommand(
-    operation="node-get",
-    input_model=NodeGetParams,
-    output_model=NodeGetResult,
-    render=render_node_properties,
-)
-
-NODE_SET_COMMAND: HeadlessCommand[NodeSetResult] = HeadlessCommand(
-    operation="node-set",
-    input_model=NodeSetParams,
-    output_model=NodeSetResult,
-    render=render_node_set,
-)
-
-NODE_REMOVE_COMMAND: HeadlessCommand[NodeRemoveResult] = HeadlessCommand(
-    operation="node-remove",
-    input_model=NodeRemoveParams,
-    output_model=NodeRemoveResult,
-    render=render_node_remove,
-)
-
-NODE_DUPLICATE_COMMAND: HeadlessCommand[NodeDuplicateResult] = HeadlessCommand(
-    operation="node-duplicate",
-    input_model=NodeDuplicateParams,
-    output_model=NodeDuplicateResult,
-    render=render_node_duplicate,
-)
-
-NODE_MOVE_COMMAND: HeadlessCommand[NodeMoveResult] = HeadlessCommand(
-    operation="node-move",
-    input_model=NodeMoveParams,
-    output_model=NodeMoveResult,
-    render=render_node_move,
-)
-
-NODE_CONNECT_SIGNAL_COMMAND: HeadlessCommand[NodeConnectSignalResult] = HeadlessCommand(
-    operation="node-connect-signal",
-    input_model=NodeConnectSignalParams,
-    output_model=NodeConnectSignalResult,
-    render=render_node_connect_signal,
-)
-
-NODE_DISCONNECT_SIGNAL_COMMAND: HeadlessCommand[NodeDisconnectSignalResult] = (
-    HeadlessCommand(
-        operation="node-disconnect-signal",
-        input_model=NodeDisconnectSignalParams,
-        output_model=NodeDisconnectSignalResult,
-        render=render_node_disconnect_signal,
-    )
-)
-
 SCRIPT_CREATE_COMMAND: HeadlessCommand[ScriptCreateResult] = HeadlessCommand(
     operation="script-create",
     input_model=ScriptCreateParams,
@@ -1887,470 +1741,6 @@ PROJECT_STATISTICS_COMMAND: HeadlessCommand[ProjectStatisticsResult] = HeadlessC
 # command's body (``export run`` included, since ADR-0023 routed it through a built
 # ``ExportRunParams``) passes its raw path straight to the params model, which
 # ~-expands it. There is no CLI-layer normalization step left to share.
-
-
-@scene_app.command(cls=SCENE_CREATE_COMMAND.command_class())
-def create(
-    path: str = typer.Argument(..., help="Target .tscn path to write."),
-    root_type: str = typer.Option(
-        ...,
-        "--root-type",
-        help="Godot node class of the new scene's root (e.g. Node2D).",
-    ),
-    root_name: Optional[str] = typer.Option(
-        None,
-        "--root-name",
-        help=(
-            "Root node name to write. Defaults to the target filename without "
-            "its final extension."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = SCENE_CREATE_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Create a new .tscn scene file with the given root node type."""
-    # Normalization + root-name derivation live in SceneCreateParams (ADR-0015),
-    # so this body is a thin argv→model adapter and the --params-json path agrees.
-    _dispatch(
-        SCENE_CREATE_COMMAND,
-        SceneCreateParams(path=path, root_type=root_type, root_name=root_name),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@scene_app.command(cls=SCENE_GET_COMMAND.command_class())
-def get(
-    path: str = typer.Argument(..., help="The .tscn scene file to read."),
-    json_output: bool = json_option(),
-    schema: bool = SCENE_GET_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Read a scene file and report its structured node tree."""
-    _dispatch(
-        SCENE_GET_COMMAND,
-        SceneGetParams(path=path),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@scene_app.command(name="get-exports", cls=SCENE_GET_EXPORTS_COMMAND.command_class())
-def get_exports(
-    path: str = typer.Argument(..., help="The .tscn scene file to read."),
-    json_output: bool = json_option(),
-    schema: bool = SCENE_GET_EXPORTS_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """List the @export properties a scene's nodes' scripts declare, per node path."""
-    _dispatch(
-        SCENE_GET_EXPORTS_COMMAND,
-        SceneGetExportsParams(path=path),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@scene_app.command(name="list", cls=SCENE_LIST_COMMAND.command_class())
-def list_scenes(
-    json_output: bool = json_option(),
-    schema: bool = SCENE_LIST_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Enumerate the .tscn scenes in the resolved project."""
-    _dispatch(
-        SCENE_LIST_COMMAND,
-        SceneListParams(),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@scene_app.command(cls=SCENE_DELETE_COMMAND.command_class())
-def delete(
-    path: str = typer.Argument(..., help="The .tscn scene file to delete."),
-    json_output: bool = json_option(),
-    schema: bool = SCENE_DELETE_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Delete a scene file and report what was removed."""
-    _dispatch(
-        SCENE_DELETE_COMMAND,
-        SceneDeleteParams(path=path),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(cls=NODE_ADD_COMMAND.command_class())
-def add(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    node_type: Optional[str] = typer.Option(
-        None,
-        "--type",
-        help=(
-            "Node type to add: a Godot node class (e.g. Sprite2D), or a "
-            "class_name registered in the project's global class list. "
-            "Exactly one of --type/--instance must be given."
-        ),
-    ),
-    instance: Optional[str] = typer.Option(
-        None,
-        "--instance",
-        help=(
-            "Scene file to add as an instanced child (e.g. res://hud.tscn). "
-            "Exactly one of --type/--instance must be given."
-        ),
-    ),
-    parent: str = typer.Option(
-        ".",
-        "--parent",
-        help=(
-            "Parent node path, relative to the scene root: '.' addresses the "
-            "root itself, 'Player/Arm' a nested node."
-        ),
-    ),
-    name: Optional[str] = typer.Option(
-        None,
-        "--name",
-        help=(
-            "Name for the new node. Defaults to the type name, or the "
-            "instanced scene's filename stem."
-        ),
-    ),
-    index: Optional[int] = typer.Option(
-        None,
-        "--index",
-        help=(
-            "0-based sibling index under the parent where the child is inserted. "
-            "Omit to append; child_count appends."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_ADD_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Add a node to a scene file under the given parent node path."""
-    # The exactly-one-of --type/--instance rule lives on the model (ADR-0015);
-    # the argv path keeps usage-error ergonomics (exit 2), --params-json
-    # surfaces the same rule as a structured invalid_params.
-    try:
-        params = NodeAddParams(
-            path=path,
-            parent=parent,
-            type=node_type,
-            instance=instance,
-            name=name,
-            index=index,
-        )
-    except (ValueError, ValidationError) as exc:
-        raise typer.BadParameter(str(exc)) from exc
-    _dispatch(
-        NODE_ADD_COMMAND,
-        params,
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(name="list", cls=NODE_LIST_COMMAND.command_class())
-def list_nodes(
-    path: str = typer.Argument(..., help="The .tscn scene file to read."),
-    json_output: bool = json_option(),
-    schema: bool = NODE_LIST_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """List a scene's node tree with each node's path relative to the root."""
-    _dispatch(
-        NODE_LIST_COMMAND,
-        NodeListParams(path=path),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(cls=NODE_GET_COMMAND.command_class())
-def get(
-    path: str = typer.Argument(..., help="The .tscn scene file to read."),
-    node: str = typer.Option(
-        ...,
-        "--node",
-        help=(
-            "Node path, relative to the scene root: '.' addresses the root "
-            "itself, 'Player/Arm' a nested node."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_GET_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Read a node's properties (by node path) as typed JSON."""
-    _dispatch(
-        NODE_GET_COMMAND,
-        NodeGetParams(path=path, node=node),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(name="set", cls=NODE_SET_COMMAND.command_class())
-def set_property(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    node: str = typer.Option(
-        ...,
-        "--node",
-        help=(
-            "Node path, relative to the scene root: '.' addresses the root "
-            "itself, 'Player/Arm' a nested node."
-        ),
-    ),
-    property: str = typer.Option(
-        ..., "--property", help="The property to set (e.g. position, visible)."
-    ),
-    value: str = typer.Option(
-        ...,
-        "--value",
-        help=(
-            "The value to set, as a string. Coerced to the property's declared "
-            "Godot type: Vector2/Vector2i/Color take comma-separated components "
-            '(e.g. "48,72", "0.2,0.6,1,1"), and a property expecting a Resource '
-            "(sub)class takes a res:// path to an existing Resource of that class. "
-            "An uncoercible value is a clean error."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_SET_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Set a node property, coercing the value to its declared Godot type."""
-    _dispatch(
-        NODE_SET_COMMAND,
-        NodeSetParams(path=path, node=node, property=property, value=value),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(name="remove", cls=NODE_REMOVE_COMMAND.command_class())
-def remove_node(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    node: str = typer.Option(
-        ...,
-        "--node",
-        help=(
-            "Node path of the node to delete, relative to the scene root: "
-            "'Player/Arm' a nested node. The root ('.') cannot be removed."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_REMOVE_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Remove a node (and its subtree) from a scene file by node path."""
-    _dispatch(
-        NODE_REMOVE_COMMAND,
-        NodeRemoveParams(path=path, node=node),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(name="duplicate", cls=NODE_DUPLICATE_COMMAND.command_class())
-def duplicate_node(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    node: str = typer.Option(
-        ...,
-        "--node",
-        help=(
-            "Node path of the node to copy, relative to the scene root: "
-            "'Player/Arm' a nested node. The copy lands under this node's own "
-            "parent with a fresh name. The root ('.') cannot be duplicated."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_DUPLICATE_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Duplicate a node (and its subtree) under its parent with a fresh name."""
-    _dispatch(
-        NODE_DUPLICATE_COMMAND,
-        NodeDuplicateParams(path=path, node=node),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(name="move", cls=NODE_MOVE_COMMAND.command_class())
-def move_node(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    node: str = typer.Option(
-        ...,
-        "--node",
-        help=(
-            "Node path of the node to reparent, relative to the scene root: "
-            "'Player/Arm' a nested node. The root ('.') cannot be moved."
-        ),
-    ),
-    to: str = typer.Option(
-        ...,
-        "--to",
-        help=(
-            "Node path of the new parent, relative to the scene root: '.' "
-            "addresses the root itself, 'Enemies' a nested node. Must not be the "
-            "moved node or one of its descendants (a cyclic target)."
-        ),
-    ),
-    index: Optional[int] = typer.Option(
-        None,
-        "--index",
-        help=(
-            "Final 0-based sibling index under --to. Omit to append on "
-            "cross-parent moves and no-op on same-parent moves."
-        ),
-    ),
-    json_output: bool = json_option(),
-    schema: bool = NODE_MOVE_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Reparent a node (and its subtree) under a new parent node path."""
-    _dispatch(
-        NODE_MOVE_COMMAND,
-        NodeMoveParams(path=path, node=node, to=to, index=index),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-# The four connection flags reused by both connect-signal and disconnect-signal.
-# Defined once so the source/target node-path addressing and the signal/method
-# naming stay identical across the wire and unwire commands.
-def _from_option() -> str:
-    return typer.Option(
-        ...,
-        "--from",
-        help=(
-            "Source node path, relative to the scene root: '.' addresses the "
-            "root itself, 'Player/Arm' a nested node."
-        ),
-    )
-
-
-def _signal_option() -> str:
-    return typer.Option(..., "--signal", help="The signal name on the source node.")
-
-
-def _to_option() -> str:
-    return typer.Option(
-        ...,
-        "--to",
-        help=(
-            "Target node path, relative to the scene root: '.' addresses the "
-            "root itself, 'Player/Arm' a nested node."
-        ),
-    )
-
-
-def _method_option() -> str:
-    return typer.Option(..., "--method", help="The method name on the target node.")
-
-
-@node_app.command(
-    name="connect-signal", cls=NODE_CONNECT_SIGNAL_COMMAND.command_class()
-)
-def connect_signal(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    from_node: str = _from_option(),
-    signal: str = _signal_option(),
-    to: str = _to_option(),
-    method: str = _method_option(),
-    json_output: bool = json_option(),
-    schema: bool = NODE_CONNECT_SIGNAL_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Wire a source node's signal to a target node's method, persisted in the scene."""
-    _dispatch(
-        NODE_CONNECT_SIGNAL_COMMAND,
-        NodeConnectSignalParams(
-            path=path,
-            from_node=from_node,
-            signal=signal,
-            to=to,
-            method=method,
-        ),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
-
-
-@node_app.command(
-    name="disconnect-signal", cls=NODE_DISCONNECT_SIGNAL_COMMAND.command_class()
-)
-def disconnect_signal(
-    path: str = typer.Argument(..., help="The .tscn scene file to mutate."),
-    from_node: str = _from_option(),
-    signal: str = _signal_option(),
-    to: str = _to_option(),
-    method: str = _method_option(),
-    json_output: bool = json_option(),
-    schema: bool = NODE_DISCONNECT_SIGNAL_COMMAND.schema_option(),
-    params_json: Optional[str] = params_json_option(),
-    godot: Optional[str] = godot_option(),
-    project: Optional[str] = project_option(),
-) -> None:
-    """Unwire an existing signal→method connection; errors if it is absent."""
-    _dispatch(
-        NODE_DISCONNECT_SIGNAL_COMMAND,
-        NodeDisconnectSignalParams(
-            path=path,
-            from_node=from_node,
-            signal=signal,
-            to=to,
-            method=method,
-        ),
-        json_output=json_output,
-        godot=godot,
-        project=project,
-    )
 
 
 @script_app.command(cls=SCRIPT_CREATE_COMMAND.command_class())
