+++ src/gda/commands/scene.py
@@ -1,0 +1,518 @@
+"""The ``scene`` command group: Godot scene files (.tscn) as the domain object.
+
+One vertical slice per `Command group` (ADR-0040): this module owns the group's
+params/result models, its human renderers, its ``HeadlessCommand`` descriptors
+(ADR-0023), and its Typer command bodies, and mounts them on the root app
+through :func:`register`. It imports the shared machinery downward — the
+dispatch tail (``gda.dispatch``), the descriptor machinery (``gda.headless``),
+the cross-command contract core (``gda.models``) and the shared render helpers
+(``gda.render``) — and is imported by nothing but the composition root
+(``gda.cli``) and its one sanctioned sibling, ``gda.commands.node`` (which
+reuses ``SceneNode`` / ``derive_scene_root_name``, ADR-0040 §5).
+"""
+
+from enum import Enum
+from typing import Any, Optional
+
+import typer
+from pydantic import BaseModel, Field, model_validator
+
+from gda.dispatch import _dispatch
+from gda.headless import (
+    HeadlessCommand,
+    godot_option,
+    json_option,
+    params_json_option,
+    project_option,
+)
+from gda.models import (
+    NormalizedPath,
+    _projected_value_schema_extra,
+    _VALUE_PROJECTION_DESC,
+)
+from gda.render import format_value, render_node_tree
+
+
+def derive_scene_root_name(path: str) -> str:
+    """Derive the default scene root name from the target file name."""
+    filename = path.replace("\\", "/").rstrip("/").rsplit("/", 1)[-1]
+    if "." in filename:
+        return filename.rsplit(".", 1)[0]
+    return filename
+
+
+class SceneCreateParams(BaseModel):
+    """The operation params of ``gda scene create`` (issue #18).
+
+    ``path`` is the target ``.tscn`` file; ``root_type`` the Godot node class
+    of the new scene's root (e.g. ``Node2D``). ``root_name`` is explicit so the
+    operation never silently derives a name Godot later sanitizes; when omitted,
+    it is derived from the target filename without the final extension. Path
+    normalization and that derivation live in the model (ADR-0015), so the argv
+    and ``--params-json`` paths produce identical params.
+    """
+
+    path: NormalizedPath = Field(description="Target .tscn path to write.")
+    root_type: str = Field(
+        description="Godot node class of the new scene's root (e.g. Node2D)."
+    )
+    root_name: str | None = Field(
+        default=None,
+        description=(
+            "Root node name to write. If omitted, it is derived from the target "
+            "filename without its final extension. Must be non-empty and must not "
+            "contain '.', ':', '@', '/', '\"', or '%'."
+        ),
+    )
+
+    @model_validator(mode="after")
+    def _default_root_name(self) -> "SceneCreateParams":
+        if self.root_name is None:
+            self.root_name = derive_scene_root_name(self.path)
+        return self
+
+
+class SceneCreateResult(BaseModel):
+    """The result of ``gda scene create``: what was written where.
+
+    Echoes the saved path and the root node the operation actually created, so
+    an agent can assert the effect without a second call. ``created_dirs`` lists
+    parent directories the operation created before saving, from outermost to
+    innermost.
+    """
+
+    path: str
+    root_name: str
+    root_type: str
+    created_dirs: list[str] = Field(
+        description=(
+            "Parent directories created before saving, from outermost to innermost."
+        )
+    )
+
+
+class SceneInstanceStatus(str, Enum):
+    """Whether a statically-read instanced scene reference resolved."""
+
+    RESOLVED = "resolved"
+    MISSING = "missing"
+
+
+class SceneNode(BaseModel):
+    """One node of a scene's structured tree: name, type, instance marker, children.
+
+    Recursive on purpose — the tree IS the contract: ``gda scene get`` reports
+    arbitrarily nested scenes through this one shape.
+    """
+
+    name: str
+    type: str = Field(
+        description=(
+            "Godot node class. For an instanced scene node, this is the "
+            "instanced scene's root class when it can be resolved statically."
+        )
+    )
+    instance_path: str | None = Field(
+        default=None,
+        exclude_if=lambda value: value is None,
+        description=(
+            "The referenced PackedScene path when this node is an instanced "
+            "scene; null for a plain typed node."
+        ),
+    )
+    instance_status: SceneInstanceStatus | None = Field(
+        default=None,
+        exclude_if=lambda value: value is None,
+        description=(
+            "Whether the instanced scene reference resolved. Null for a plain "
+            "typed node; 'missing' means instance_path is visible but could not "
+            "be loaded as a PackedScene."
+        ),
+    )
+    children: list["SceneNode"] = []
+
+
+class SceneGetParams(BaseModel):
+    """The operation params of ``gda scene get``: the ``.tscn`` file to read."""
+
+    path: NormalizedPath = Field(description="The .tscn scene file to read.")
+
+
+class SceneGetResult(BaseModel):
+    """The result of ``gda scene get``: the scene file's structured node tree."""
+
+    path: str
+    root: SceneNode
+
+
+class SceneExport(BaseModel):
+    """One ``@export`` property a node's attached script declares (issue #58).
+
+    ``type`` is the property's declared Godot type name (``float``, ``String``,
+    ``Vector2``, …), the same spelling :class:`NodeProperty` uses. ``hint`` is the
+    Godot ``PropertyHint`` enum value the ``@export`` annotation produced (e.g. a
+    ``@export_range`` yields ``PROPERTY_HINT_RANGE``); ``hint_string`` is its
+    companion string (the range bounds, the enum members, the file filter, …) —
+    together they capture HOW the export is meant to be edited. ``value`` is the
+    property's current value in the same recursive JSON value projection
+    ``node get`` reports (ADR-0035), which on a freshly-instantiated node is
+    the export's default.
+    """
+
+    name: str
+    type: str = Field(
+        description="The export's declared Godot type name (e.g. float, String, Vector2)."
+    )
+    hint: int = Field(
+        description=(
+            "The Godot PropertyHint enum value the @export annotation produced "
+            "(0 = PROPERTY_HINT_NONE)."
+        )
+    )
+    hint_string: str = Field(
+        description="The PropertyHint's companion string (range bounds, enum members, …); empty when none."
+    )
+    value: Any = Field(
+        description=(
+            "The export's current value as JSON (its default on a freshly-loaded "
+            "node). " + _VALUE_PROJECTION_DESC
+        ),
+        json_schema_extra=_projected_value_schema_extra,
+    )
+
+
+class ExportingNode(BaseModel):
+    """One node of ``gda scene get-exports``: the exports its script declares (issue #58).
+
+    A node appears only when its attached script declares at least one ``@export``
+    property. ``path`` is the node's node path relative to the scene root ('.' for
+    the root), the same addressing ``node get`` uses, so an agent can read or set
+    any export with ``node get``/``node set``. ``script`` is the attached script's
+    ``res://`` path, naming where the exports came from. ``exports`` lists them in
+    declaration order.
+    """
+
+    path: str = Field(
+        description=(
+            "The node's node path relative to the scene root: '.' for the root "
+            "itself, 'Player/Arm' for a nested node."
+        )
+    )
+    name: str
+    type: str = Field(description="The node's engine class (e.g. Node2D).")
+    script: str | None = Field(
+        default=None,
+        description=(
+            "The res:// path of the script that declares these exports, or null "
+            "when the attached script has no resource path (an embedded script)."
+        ),
+    )
+    exports: list[SceneExport]
+
+
+class SceneGetExportsParams(BaseModel):
+    """The operation params of ``gda scene get-exports``: the ``.tscn`` file to read (issue #58)."""
+
+    path: NormalizedPath = Field(description="The .tscn scene file to read.")
+
+
+class SceneGetExportsResult(BaseModel):
+    """The result of ``gda scene get-exports``: each node's declared ``@export``s (issue #58).
+
+    Echoes the scene ``path`` and, per node that declares them, the ``@export``
+    properties its attached script exposes — name, type, hint/hint_string, and
+    current value — as typed JSON. Reuses ``node get``'s property-value
+    projection, so an export's ``value`` reads exactly as ``node get`` would
+    report it. Nodes without an export-declaring script are omitted; a scene with
+    no exported variables anywhere is a valid, empty listing (``nodes == []``).
+    """
+
+    path: str
+    nodes: list[ExportingNode]
+
+
+class SceneListParams(BaseModel):
+    """The operation params of ``gda scene list`` — none (ADR-0004).
+
+    ``scene list`` enumerates the ``.tscn`` scenes in the resolved project's
+    ``res://`` tree; the project is process context (``--project``), not an
+    operation param (ADR-0006), so the ``input`` schema is trivially empty.
+    """
+
+
+class ListedScene(BaseModel):
+    """One enumerated scene of ``gda scene list``: its path and root summary.
+
+    ``path`` is the scene's ``res://`` path — the address an agent feeds back
+    into other scene commands. ``root_name``/``root_type`` are read cheaply from
+    the scene's stored state (no instantiation, issue #30); both are null when
+    the ``.tscn`` could not be loaded as a scene, so the entry still names a file
+    the listing found rather than dropping it.
+    """
+
+    path: str
+    root_name: str | None = Field(
+        default=None,
+        description="The scene root node's name, or null if the file could not be loaded as a scene.",
+    )
+    root_type: str | None = Field(
+        default=None,
+        description=(
+            "The scene root node's type, resolving an inherited/instanced root "
+            "to the referenced scene's root type when possible; null if the "
+            "file could not be loaded as a scene."
+        ),
+    )
+    root_instance_path: str | None = Field(
+        default=None,
+        exclude_if=lambda value: value is None,
+        description=(
+            "The referenced PackedScene path when the scene root inherits or "
+            "instances another scene; null for a plain typed root or an "
+            "unloadable scene."
+        ),
+    )
+    root_instance_status: SceneInstanceStatus | None = Field(
+        default=None,
+        exclude_if=lambda value: value is None,
+        description=(
+            "Whether the root instance reference resolved. Null for a plain "
+            "typed root or an unloadable scene; 'missing' means "
+            "root_instance_path is visible but could not be loaded as a "
+            "PackedScene."
+        ),
+    )
+
+
+class SceneListResult(BaseModel):
+    """The result of ``gda scene list``: the project's enumerated ``.tscn`` scenes.
+
+    An empty project is a valid, empty listing — ``scenes == []`` — not a
+    failure.
+    """
+
+    scenes: list[ListedScene]
+
+
+class SceneDeleteParams(BaseModel):
+    """The operation params of ``gda scene delete``: the ``.tscn`` file to remove."""
+
+    path: NormalizedPath = Field(description="The .tscn scene file to delete.")
+
+
+class SceneDeleteResult(BaseModel):
+    """The result of ``gda scene delete``: what was removed.
+
+    Echoes the deleted scene's path and its root node's name/type (read from the
+    scene's stored state before deletion), so the result names the content
+    removed, not just the file path.
+    """
+
+    path: str
+    root_name: str
+    root_type: str
+
+
+def render_scene_metadata(scene: "SceneCreateResult") -> str:
+    """Render a created scene as ``created <path> (root <type>)``."""
+    return f"created {scene.path} (root {scene.root_type})"
+
+
+def render_scene_tree(scene: "SceneGetResult") -> str:
+    """Render a read scene's node tree."""
+    return render_node_tree(scene.root)
+
+
+def render_scene_exports(scene: "SceneGetExportsResult") -> str:
+    """Render a scene's per-node @export properties for humans.
+
+    One ``path (Type)`` header per node that declares exports, then a
+    ``name (Type) = value`` line per export — reusing :func:`format_value` for
+    the value, the same projection ``node get`` renders. An empty listing (no
+    exported variables anywhere) reads as ``(no exports)``.
+    """
+    if not scene.nodes:
+        return "(no exports)"
+    lines = []
+    for node in scene.nodes:
+        lines.append(f"{node.path} ({node.type})")
+        for export in node.exports:
+            lines.append(
+                f"  {export.name} ({export.type}) = {format_value(export.value)}"
+            )
+    return "\n".join(lines)
+
+
+def render_scene_list(listed: "SceneListResult") -> str:
+    """Render the enumerated scenes as ``path (root_name: root_type)`` lines."""
+    if not listed.scenes:
+        return "(no scenes)"
+    lines = []
+    for scene in listed.scenes:
+        if scene.root_name is not None and scene.root_type is not None:
+            lines.append(f"{scene.path} ({scene.root_name}: {scene.root_type})")
+        else:
+            lines.append(f"{scene.path} (unreadable)")
+    return "\n".join(lines)
+
+
+def render_scene_delete(removed: "SceneDeleteResult") -> str:
+    """Render a deleted scene as ``deleted <path> (root <name>: <type>)``."""
+    return f"deleted {removed.path} (root {removed.root_name}: {removed.root_type})"
+
+
+SCENE_CREATE_COMMAND: HeadlessCommand[SceneCreateResult] = HeadlessCommand(
+    operation="scene-create",
+    input_model=SceneCreateParams,
+    output_model=SceneCreateResult,
+    render=render_scene_metadata,
+)
+
+SCENE_GET_COMMAND: HeadlessCommand[SceneGetResult] = HeadlessCommand(
+    operation="scene-get",
+    input_model=SceneGetParams,
+    output_model=SceneGetResult,
+    render=render_scene_tree,
+)
+
+SCENE_GET_EXPORTS_COMMAND: HeadlessCommand[SceneGetExportsResult] = HeadlessCommand(
+    operation="scene-get-exports",
+    input_model=SceneGetExportsParams,
+    output_model=SceneGetExportsResult,
+    render=render_scene_exports,
+)
+
+SCENE_LIST_COMMAND: HeadlessCommand[SceneListResult] = HeadlessCommand(
+    operation="scene-list",
+    input_model=SceneListParams,
+    output_model=SceneListResult,
+    render=render_scene_list,
+)
+
+SCENE_DELETE_COMMAND: HeadlessCommand[SceneDeleteResult] = HeadlessCommand(
+    operation="scene-delete",
+    input_model=SceneDeleteParams,
+    output_model=SceneDeleteResult,
+    render=render_scene_delete,
+)
+
+# The first domain command group (ADR-0005): commands acting on scene files.
+_app = typer.Typer(help="Act on Godot scene files (.tscn).", no_args_is_help=True)
+
+
+@_app.command(cls=SCENE_CREATE_COMMAND.command_class())
+def create(
+    path: str = typer.Argument(..., help="Target .tscn path to write."),
+    root_type: str = typer.Option(
+        ...,
+        "--root-type",
+        help="Godot node class of the new scene's root (e.g. Node2D).",
+    ),
+    root_name: Optional[str] = typer.Option(
+        None,
+        "--root-name",
+        help=(
+            "Root node name to write. Defaults to the target filename without "
+            "its final extension."
+        ),
+    ),
+    json_output: bool = json_option(),
+    schema: bool = SCENE_CREATE_COMMAND.schema_option(),
+    params_json: Optional[str] = params_json_option(),
+    godot: Optional[str] = godot_option(),
+    project: Optional[str] = project_option(),
+) -> None:
+    """Create a new .tscn scene file with the given root node type."""
+    # Normalization + root-name derivation live in SceneCreateParams (ADR-0015),
+    # so this body is a thin argv→model adapter and the --params-json path agrees.
+    _dispatch(
+        SCENE_CREATE_COMMAND,
+        SceneCreateParams(path=path, root_type=root_type, root_name=root_name),
+        json_output=json_output,
+        godot=godot,
+        project=project,
+    )
+
+
+@_app.command(cls=SCENE_GET_COMMAND.command_class())
+def get(
+    path: str = typer.Argument(..., help="The .tscn scene file to read."),
+    json_output: bool = json_option(),
+    schema: bool = SCENE_GET_COMMAND.schema_option(),
+    params_json: Optional[str] = params_json_option(),
+    godot: Optional[str] = godot_option(),
+    project: Optional[str] = project_option(),
+) -> None:
+    """Read a scene file and report its structured node tree."""
+    _dispatch(
+        SCENE_GET_COMMAND,
+        SceneGetParams(path=path),
+        json_output=json_output,
+        godot=godot,
+        project=project,
+    )
+
+
+@_app.command(name="get-exports", cls=SCENE_GET_EXPORTS_COMMAND.command_class())
+def get_exports(
+    path: str = typer.Argument(..., help="The .tscn scene file to read."),
+    json_output: bool = json_option(),
+    schema: bool = SCENE_GET_EXPORTS_COMMAND.schema_option(),
+    params_json: Optional[str] = params_json_option(),
+    godot: Optional[str] = godot_option(),
+    project: Optional[str] = project_option(),
+) -> None:
+    """List the @export properties a scene's nodes' scripts declare, per node path."""
+    _dispatch(
+        SCENE_GET_EXPORTS_COMMAND,
+        SceneGetExportsParams(path=path),
+        json_output=json_output,
+        godot=godot,
+        project=project,
+    )
+
+
+@_app.command(name="list", cls=SCENE_LIST_COMMAND.command_class())
+def list_scenes(
+    json_output: bool = json_option(),
+    schema: bool = SCENE_LIST_COMMAND.schema_option(),
+    params_json: Optional[str] = params_json_option(),
+    godot: Optional[str] = godot_option(),
+    project: Optional[str] = project_option(),
+) -> None:
+    """Enumerate the .tscn scenes in the resolved project."""
+    _dispatch(
+        SCENE_LIST_COMMAND,
+        SceneListParams(),
+        json_output=json_output,
+        godot=godot,
+        project=project,
+    )
+
+
+@_app.command(cls=SCENE_DELETE_COMMAND.command_class())
+def delete(
+    path: str = typer.Argument(..., help="The .tscn scene file to delete."),
+    json_output: bool = json_option(),
+    schema: bool = SCENE_DELETE_COMMAND.schema_option(),
+    params_json: Optional[str] = params_json_option(),
+    godot: Optional[str] = godot_option(),
+    project: Optional[str] = project_option(),
+) -> None:
+    """Delete a scene file and report what was removed."""
+    _dispatch(
+        SCENE_DELETE_COMMAND,
+        SceneDeleteParams(path=path),
+        json_output=json_output,
+        godot=godot,
+        project=project,
+    )
+
+
+def register(root: typer.Typer) -> None:
+    """Mount the ``scene`` group on the root app (ADR-0040).
+
+    Mounting IS the registration: the live Typer tree stays the only registry
+    (ADR-0012/0023), so no parallel table records this group.
+    """
+    root.add_typer(_app, name="scene")
