+++ src/gda/models.py
@@ -314,118 +314,6 @@ def normalize_export_output_path(path: str) -> str:
 ExportOutputPath = Annotated[str, AfterValidator(normalize_export_output_path)]
 
 
-def derive_scene_root_name(path: str) -> str:
-    """Derive the default scene root name from the target file name."""
-    filename = path.replace("\\", "/").rstrip("/").rsplit("/", 1)[-1]
-    if "." in filename:
-        return filename.rsplit(".", 1)[0]
-    return filename
-
-
-class SceneCreateParams(BaseModel):
-    """The operation params of ``gda scene create`` (issue #18).
-
-    ``path`` is the target ``.tscn`` file; ``root_type`` the Godot node class
-    of the new scene's root (e.g. ``Node2D``). ``root_name`` is explicit so the
-    operation never silently derives a name Godot later sanitizes; when omitted,
-    it is derived from the target filename without the final extension. Path
-    normalization and that derivation live in the model (ADR-0015), so the argv
-    and ``--params-json`` paths produce identical params.
-    """
-
-    path: NormalizedPath = Field(description="Target .tscn path to write.")
-    root_type: str = Field(
-        description="Godot node class of the new scene's root (e.g. Node2D)."
-    )
-    root_name: str | None = Field(
-        default=None,
-        description=(
-            "Root node name to write. If omitted, it is derived from the target "
-            "filename without its final extension. Must be non-empty and must not "
-            "contain '.', ':', '@', '/', '\"', or '%'."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _default_root_name(self) -> "SceneCreateParams":
-        if self.root_name is None:
-            self.root_name = derive_scene_root_name(self.path)
-        return self
-
-
-class SceneCreateResult(BaseModel):
-    """The result of ``gda scene create``: what was written where.
-
-    Echoes the saved path and the root node the operation actually created, so
-    an agent can assert the effect without a second call. ``created_dirs`` lists
-    parent directories the operation created before saving, from outermost to
-    innermost.
-    """
-
-    path: str
-    root_name: str
-    root_type: str
-    created_dirs: list[str] = Field(
-        description=(
-            "Parent directories created before saving, from outermost to innermost."
-        )
-    )
-
-
-class SceneInstanceStatus(str, Enum):
-    """Whether a statically-read instanced scene reference resolved."""
-
-    RESOLVED = "resolved"
-    MISSING = "missing"
-
-
-class SceneNode(BaseModel):
-    """One node of a scene's structured tree: name, type, instance marker, children.
-
-    Recursive on purpose — the tree IS the contract: ``gda scene get`` reports
-    arbitrarily nested scenes through this one shape.
-    """
-
-    name: str
-    type: str = Field(
-        description=(
-            "Godot node class. For an instanced scene node, this is the "
-            "instanced scene's root class when it can be resolved statically."
-        )
-    )
-    instance_path: str | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "The referenced PackedScene path when this node is an instanced "
-            "scene; null for a plain typed node."
-        ),
-    )
-    instance_status: SceneInstanceStatus | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "Whether the instanced scene reference resolved. Null for a plain "
-            "typed node; 'missing' means instance_path is visible but could not "
-            "be loaded as a PackedScene."
-        ),
-    )
-    children: list["SceneNode"] = []
-
-
-class SceneGetParams(BaseModel):
-    """The operation params of ``gda scene get``: the ``.tscn`` file to read."""
-
-    path: NormalizedPath = Field(description="The .tscn scene file to read.")
-
-
-class SceneGetResult(BaseModel):
-    """The result of ``gda scene get``: the scene file's structured node tree."""
-
-    path: str
-    root: SceneNode
-
-
 # The one read-side value projection every value gda emits goes through
 # (ADR-0035): the shared field description for the dynamically-shaped `value`
 # fields. The field itself stays `Any` — a value's shape is not statically
@@ -528,333 +416,6 @@ def _projected_value_schema_extra(schema: dict[str, Any]) -> None:
     }
 
 
-class SceneExport(BaseModel):
-    """One ``@export`` property a node's attached script declares (issue #58).
-
-    ``type`` is the property's declared Godot type name (``float``, ``String``,
-    ``Vector2``, …), the same spelling :class:`NodeProperty` uses. ``hint`` is the
-    Godot ``PropertyHint`` enum value the ``@export`` annotation produced (e.g. a
-    ``@export_range`` yields ``PROPERTY_HINT_RANGE``); ``hint_string`` is its
-    companion string (the range bounds, the enum members, the file filter, …) —
-    together they capture HOW the export is meant to be edited. ``value`` is the
-    property's current value in the same recursive JSON value projection
-    ``node get`` reports (ADR-0035), which on a freshly-instantiated node is
-    the export's default.
-    """
-
-    name: str
-    type: str = Field(
-        description="The export's declared Godot type name (e.g. float, String, Vector2)."
-    )
-    hint: int = Field(
-        description=(
-            "The Godot PropertyHint enum value the @export annotation produced "
-            "(0 = PROPERTY_HINT_NONE)."
-        )
-    )
-    hint_string: str = Field(
-        description="The PropertyHint's companion string (range bounds, enum members, …); empty when none."
-    )
-    value: Any = Field(
-        description=(
-            "The export's current value as JSON (its default on a freshly-loaded "
-            "node). " + _VALUE_PROJECTION_DESC
-        ),
-        json_schema_extra=_projected_value_schema_extra,
-    )
-
-
-class ExportingNode(BaseModel):
-    """One node of ``gda scene get-exports``: the exports its script declares (issue #58).
-
-    A node appears only when its attached script declares at least one ``@export``
-    property. ``path`` is the node's node path relative to the scene root ('.' for
-    the root), the same addressing ``node get`` uses, so an agent can read or set
-    any export with ``node get``/``node set``. ``script`` is the attached script's
-    ``res://`` path, naming where the exports came from. ``exports`` lists them in
-    declaration order.
-    """
-
-    path: str = Field(
-        description=(
-            "The node's node path relative to the scene root: '.' for the root "
-            "itself, 'Player/Arm' for a nested node."
-        )
-    )
-    name: str
-    type: str = Field(description="The node's engine class (e.g. Node2D).")
-    script: str | None = Field(
-        default=None,
-        description=(
-            "The res:// path of the script that declares these exports, or null "
-            "when the attached script has no resource path (an embedded script)."
-        ),
-    )
-    exports: list[SceneExport]
-
-
-class SceneGetExportsParams(BaseModel):
-    """The operation params of ``gda scene get-exports``: the ``.tscn`` file to read (issue #58)."""
-
-    path: NormalizedPath = Field(description="The .tscn scene file to read.")
-
-
-class SceneGetExportsResult(BaseModel):
-    """The result of ``gda scene get-exports``: each node's declared ``@export``s (issue #58).
-
-    Echoes the scene ``path`` and, per node that declares them, the ``@export``
-    properties its attached script exposes — name, type, hint/hint_string, and
-    current value — as typed JSON. Reuses ``node get``'s property-value
-    projection, so an export's ``value`` reads exactly as ``node get`` would
-    report it. Nodes without an export-declaring script are omitted; a scene with
-    no exported variables anywhere is a valid, empty listing (``nodes == []``).
-    """
-
-    path: str
-    nodes: list[ExportingNode]
-
-
-class SceneListParams(BaseModel):
-    """The operation params of ``gda scene list`` — none (ADR-0004).
-
-    ``scene list`` enumerates the ``.tscn`` scenes in the resolved project's
-    ``res://`` tree; the project is process context (``--project``), not an
-    operation param (ADR-0006), so the ``input`` schema is trivially empty.
-    """
-
-
-class ListedScene(BaseModel):
-    """One enumerated scene of ``gda scene list``: its path and root summary.
-
-    ``path`` is the scene's ``res://`` path — the address an agent feeds back
-    into other scene commands. ``root_name``/``root_type`` are read cheaply from
-    the scene's stored state (no instantiation, issue #30); both are null when
-    the ``.tscn`` could not be loaded as a scene, so the entry still names a file
-    the listing found rather than dropping it.
-    """
-
-    path: str
-    root_name: str | None = Field(
-        default=None,
-        description="The scene root node's name, or null if the file could not be loaded as a scene.",
-    )
-    root_type: str | None = Field(
-        default=None,
-        description=(
-            "The scene root node's type, resolving an inherited/instanced root "
-            "to the referenced scene's root type when possible; null if the "
-            "file could not be loaded as a scene."
-        ),
-    )
-    root_instance_path: str | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "The referenced PackedScene path when the scene root inherits or "
-            "instances another scene; null for a plain typed root or an "
-            "unloadable scene."
-        ),
-    )
-    root_instance_status: SceneInstanceStatus | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "Whether the root instance reference resolved. Null for a plain "
-            "typed root or an unloadable scene; 'missing' means "
-            "root_instance_path is visible but could not be loaded as a "
-            "PackedScene."
-        ),
-    )
-
-
-class SceneListResult(BaseModel):
-    """The result of ``gda scene list``: the project's enumerated ``.tscn`` scenes.
-
-    An empty project is a valid, empty listing — ``scenes == []`` — not a
-    failure.
-    """
-
-    scenes: list[ListedScene]
-
-
-class SceneDeleteParams(BaseModel):
-    """The operation params of ``gda scene delete``: the ``.tscn`` file to remove."""
-
-    path: NormalizedPath = Field(description="The .tscn scene file to delete.")
-
-
-class SceneDeleteResult(BaseModel):
-    """The result of ``gda scene delete``: what was removed.
-
-    Echoes the deleted scene's path and its root node's name/type (read from the
-    scene's stored state before deletion), so the result names the content
-    removed, not just the file path.
-    """
-
-    path: str
-    root_name: str
-    root_type: str
-
-
-class NodeAddParams(BaseModel):
-    """The operation params of ``gda node add`` (issue #53; instancing #399).
-
-    ``path`` is the ``.tscn`` scene file to mutate. ``parent`` addresses the
-    parent node by node path. Exactly one of ``type``/``instance`` selects what
-    is added: ``type`` is resolved first as a built-in Godot node class, then
-    as a ``class_name`` registered in the project's global class list;
-    ``instance`` composes an existing scene as an instanced child (#399).
-    ``name`` is explicit so the operation never silently derives a name Godot
-    later sanitizes; when the CLI caller omits ``--name``, it uses the type
-    name, or the instanced scene's filename stem.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    parent: str = Field(
-        default=".",
-        description=(
-            "Parent node path, relative to the scene root: '.' addresses the "
-            "root itself, 'Player/Arm' a nested node."
-        ),
-    )
-    type: str | None = Field(
-        default=None,
-        description=(
-            "Node type to add: a Godot node class (e.g. Sprite2D), or a "
-            "class_name registered in the project's global class list. "
-            "Exactly one of type/instance must be given."
-        ),
-    )
-    instance: NormalizedPath | None = Field(
-        default=None,
-        description=(
-            "Scene file to add as an instanced child (e.g. res://hud.tscn) — "
-            "composes the scene under the parent, serialized as an "
-            "instance=ExtResource(...) entry. Exactly one of type/instance "
-            "must be given."
-        ),
-    )
-    name: str | None = Field(
-        default=None,
-        description=(
-            "Name for the new node. If omitted, the type name (or the "
-            "instanced scene's filename stem) is used. Must be non-empty and "
-            "must not contain '.', ':', '@', '/', '\"', or '%'."
-        ),
-    )
-    index: int | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "Optional 0-based sibling index under the parent where the new "
-            "child is inserted. Omit to append. Valid runtime range is "
-            "0..child_count before insertion, so child_count appends."
-        ),
-    )
-
-    @model_validator(mode="after")
-    def _exactly_one_mode_and_default_name(self) -> "NodeAddParams":
-        # Exactly one of type/instance selects what is added (#399). Enforced
-        # model-side (ADR-0015) so the argv and --params-json paths agree: the
-        # argv path converts the ValueError to a usage error, --params-json
-        # surfaces it as a structured invalid_params.
-        if self.type is None and self.instance is None:
-            raise ValueError(
-                "node add needs exactly one of --type or --instance "
-                "(neither was given)."
-            )
-        if self.type is not None and self.instance is not None:
-            raise ValueError(
-                "--type and --instance are mutually exclusive; pass exactly one."
-            )
-        # Derive the default node name model-side too, so the CLI never
-        # derives it: the type name, or the instanced scene's filename stem.
-        if self.name is None:
-            self.name = (
-                self.type
-                if self.type is not None
-                else derive_scene_root_name(self.instance or "")
-            )
-        return self
-
-
-class NodeAddResult(BaseModel):
-    """The result of ``gda node add``: the created node and where it landed.
-
-    ``path`` is the created node's node path relative to the scene root, so an
-    agent can address the node in follow-up node commands without re-listing.
-    """
-
-    scene_path: str
-    path: str = Field(
-        description="The created node's node path, relative to the scene root."
-    )
-    name: str
-    type: str = Field(
-        description=(
-            "The created node's engine class (e.g. Node2D) — for a class_name "
-            "addition, the script's base class."
-        )
-    )
-    script_class: str | None = Field(
-        default=None,
-        description=(
-            "The class_name of the script attached to the created node, when "
-            "the requested type resolved to a script class; null for a "
-            "built-in type."
-        ),
-    )
-    instance: str | None = Field(
-        default=None,
-        description=(
-            "The res:// path of the scene this node instances, when the node "
-            "was added via --instance (#399); null for a type addition. For "
-            "an instance, `type` reports the instanced scene's root class."
-        ),
-    )
-
-
-class ListedNode(SceneNode):
-    """One node of ``gda node list``'s tree: name, type, node path, children.
-
-    Like ``SceneNode`` but each node also carries its node path relative to the
-    scene root ('.' for the root itself) — the address an agent feeds back into
-    other node commands (e.g. ``node add --parent``).
-    """
-
-    # Subclasses ``SceneNode`` so the recursive tree shape (name/type/children)
-    # has one source of truth; ``children`` is re-declared as ``list[ListedNode]``
-    # — not the inherited ``list[SceneNode]`` — so a listed node's children are
-    # themselves listed nodes (carrying ``path``), keeping the tree self-recursive
-    # on ``ListedNode``. The ``--json``/``--schema`` output stays SEMANTICALLY
-    # unchanged (same fields, same ``required``, same recursion) but is NOT byte-
-    # identical: inheriting name/type/children and appending ``path`` orders the
-    # properties name,type,children,path rather than the prior name,type,path,
-    # children. JSON property order is insignificant — consumers parse by key and
-    # the ``--schema`` → MCP generation is order-insensitive — so this is a safe
-    # trade for the single-source-of-truth tree shape.
-    path: str = Field(
-        description=(
-            "The node's node path relative to the scene root: '.' for the root "
-            "itself, 'Player/Arm' for a nested node."
-        )
-    )
-    children: list["ListedNode"] = []
-
-
-class NodeListParams(BaseModel):
-    """The operation params of ``gda node list``: the ``.tscn`` file to read."""
-
-    path: NormalizedPath = Field(description="The .tscn scene file to read.")
-
-
-class NodeListResult(BaseModel):
-    """The result of ``gda node list``: the scene's node tree with node paths."""
-
-    scene_path: str
-    root: ListedNode
-
-
 class NodeProperty(BaseModel):
     """One of a node's properties as ``gda node get`` reports it (issue #55).
 
@@ -865,6 +426,10 @@ class NodeProperty(BaseModel):
     becomes ``[x, y]``, a Dictionary a JSON object, an Object a
     :class:`ReferenceProjection` / :class:`InlineValueProjection` / ``str()``
     fallback.
+
+    Stays in the shared core (ADR-0040 §4): three groups read the SAME typed
+    property shape — ``node get``, ``resource get`` and the live ``game get`` —
+    so it is a cross-command contract, not one group's model.
     """
 
     name: str
@@ -875,334 +440,6 @@ class NodeProperty(BaseModel):
         description="The property's value as JSON. " + _VALUE_PROJECTION_DESC,
         json_schema_extra=_projected_value_schema_extra,
     )
-
-
-class NodeGetParams(BaseModel):
-    """The operation params of ``gda node get`` (issue #55).
-
-    ``path`` is the ``.tscn`` scene file to read; ``node`` addresses the node by
-    its node path relative to the scene root ('.' is the root itself).
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to read.")
-    node: str = Field(
-        description=(
-            "Node path relative to the scene root: '.' addresses the root "
-            "itself, 'Player/Arm' a nested node."
-        )
-    )
-
-
-class NodeGetResult(BaseModel):
-    """The result of ``gda node get``: a node's properties as typed JSON (issue #55).
-
-    Echoes the addressed node (``path``/``name``/``type``) and its storage
-    properties — the ones that serialize into the ``.tscn`` — each as a typed
-    :class:`NodeProperty`, so an agent reads a node's state without parsing the
-    scene file and can feed any property straight back into ``node set``.
-    """
-
-    scene_path: str
-    path: str = Field(
-        description="The addressed node's node path, relative to the scene root."
-    )
-    name: str
-    type: str = Field(description="The node's engine class (e.g. Sprite2D).")
-    properties: list[NodeProperty]
-
-
-class NodeSetParams(BaseModel):
-    """The operation params of ``gda node set`` (issue #55).
-
-    ``path`` is the ``.tscn`` scene file to mutate; ``node`` addresses the node
-    by node path relative to the scene root. ``property`` names the property to
-    set; ``value`` is the CLI string value, coerced to the property's declared
-    Godot type by the operation before the scene is re-packed and saved.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    node: str = Field(
-        description=(
-            "Node path relative to the scene root: '.' addresses the root "
-            "itself, 'Player/Arm' a nested node."
-        )
-    )
-    property: str = Field(description="The property to set (e.g. position, visible).")
-    value: str = Field(
-        description=(
-            "The value to set, as a string. The operation coerces it to the "
-            "property's declared Godot type (see the command catalog's "
-            "'Property value coercion'). For Dictionary/Array JSON values, JSON "
-            "integer literals stay int and JSON float literals stay float; typed "
-            "containers assign entries through their declared container type. An "
-            "uncoercible value is a clean error."
-        )
-    )
-
-
-class NodeSetResult(BaseModel):
-    """The result of ``gda node set``: the one property it set (issue #55).
-
-    Echoes the addressed node's ``path``, the ``property`` set, the declared
-    ``type`` the CLI value was coerced to, and the coerced ``value`` as JSON —
-    the projection ``node get`` reports, so a ``set`` round-trips through a
-    ``get`` without re-reading the file.
-    """
-
-    scene_path: str
-    path: str = Field(
-        description="The addressed node's node path, relative to the scene root."
-    )
-    property: str
-    type: str = Field(
-        description="The property's declared Godot type the value was coerced to."
-    )
-    value: Any = Field(
-        description=(
-            "The coerced value as JSON, as the node now holds it. "
-            + _OBJECT_SET_ECHO_DESC
-        ),
-        json_schema_extra=_projected_value_schema_extra,
-    )
-
-
-class NodeRemoveParams(BaseModel):
-    """The operation params of ``gda node remove`` (issue #56).
-
-    ``path`` is the ``.tscn`` scene file to mutate; ``node`` addresses the node
-    to delete by its node path relative to the scene root. The scene root ('.')
-    has no parent to be removed from, so removing it is refused rather than
-    emptying the scene.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    node: str = Field(
-        description=(
-            "Node path relative to the scene root: 'Player/Arm' a nested node. "
-            "The root ('.') cannot be removed."
-        )
-    )
-
-
-class NodeRemoveResult(BaseModel):
-    """The result of ``gda node remove``: the node and subtree it deleted (issue #56).
-
-    Echoes the removed node's ``path`` (its node path relative to the scene
-    root), ``name``, and ``type``, captured off the tree before the re-save —
-    so the result names the content removed, not just the path.
-    """
-
-    scene_path: str
-    path: str = Field(
-        description="The removed node's node path, relative to the scene root."
-    )
-    name: str
-    type: str = Field(description="The removed node's engine class (e.g. Sprite2D).")
-
-
-class NodeDuplicateParams(BaseModel):
-    """The operation params of ``gda node duplicate`` (issue #56).
-
-    ``path`` is the ``.tscn`` scene file to mutate; ``node`` addresses the node
-    to copy by its node path relative to the scene root. The copy (and its whole
-    subtree) lands under the source node's own parent with a fresh,
-    non-colliding name. The scene root ('.') has no parent to host a sibling
-    copy, so duplicating it is refused.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    node: str = Field(
-        description=(
-            "Node path relative to the scene root: 'Player/Arm' a nested node. "
-            "The copy lands under this node's own parent; the root ('.') cannot "
-            "be duplicated."
-        )
-    )
-
-
-class NodeDuplicateResult(BaseModel):
-    """The result of ``gda node duplicate``: the new copy and where it landed (issue #56).
-
-    Echoes the ``source_path`` it copied and the new copy's ``path`` (its node
-    path relative to the scene root), ``name``, and ``type`` — the fresh,
-    non-colliding name the operation assigned — so an agent can address the
-    duplicate in follow-up node commands without re-listing.
-    """
-
-    scene_path: str
-    source_path: str = Field(
-        description="The copied node's node path, relative to the scene root."
-    )
-    path: str = Field(
-        description="The new copy's node path, relative to the scene root."
-    )
-    name: str = Field(description="The fresh, non-colliding name assigned to the copy.")
-    type: str = Field(description="The copy's engine class (e.g. Sprite2D).")
-
-
-class NodeMoveParams(BaseModel):
-    """The operation params of ``gda node move`` (issue #56).
-
-    ``path`` is the ``.tscn`` scene file to mutate; ``node`` addresses the node
-    to reparent by its node path relative to the scene root; ``to`` addresses the
-    new parent the same way. The move is refused when the target is invalid (no
-    such parent, or a name collision at the destination) or **cyclic** — moving a
-    node under itself or one of its own descendants would detach the subtree from
-    the scene. The scene root ('.') has no parent to be reparented out of.
-    """
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    node: str = Field(
-        description=(
-            "Node path of the node to reparent, relative to the scene root: "
-            "'Player/Arm' a nested node. The root ('.') cannot be moved."
-        )
-    )
-    to: str = Field(
-        description=(
-            "Node path of the new parent, relative to the scene root: '.' "
-            "addresses the root itself, 'Enemies' a nested node. Must not be the "
-            "moved node itself or one of its descendants (a cyclic target)."
-        )
-    )
-    index: int | None = Field(
-        default=None,
-        exclude_if=lambda value: value is None,
-        description=(
-            "Optional final 0-based sibling index under the destination parent. "
-            "Omit to preserve existing behavior: same-parent move is a no-op, "
-            "and cross-parent move appends. With the same parent, valid runtime "
-            "range is 0..child_count-1; with a different parent, 0..target_child_count "
-            "before the move, so target_child_count appends."
-        ),
-    )
-
-
-class NodeMoveResult(BaseModel):
-    """The result of ``gda node move``: the reparented node's new home (issue #56).
-
-    Echoes the ``source_path`` it moved, the ``new_parent`` it landed under, and
-    the node's new ``path`` (its node path relative to the scene root after the
-    move), ``name``, and ``type`` — so an agent can address the moved node in
-    follow-up node commands without re-listing.
-    """
-
-    scene_path: str
-    source_path: str = Field(
-        description="The moved node's original node path, relative to the scene root."
-    )
-    new_parent: str = Field(
-        description="The new parent's node path, relative to the scene root."
-    )
-    path: str = Field(
-        description="The moved node's new node path, relative to the scene root."
-    )
-    name: str
-    type: str = Field(description="The moved node's engine class (e.g. Sprite2D).")
-
-
-# A signal→method connection's four parts (issue #57): a source node's signal
-# wired to a target node's method, the shape of a ``.tscn`` ``[connection]``.
-# ``from`` is the source node's node path and ``to`` the target's, both relative
-# to the scene root — the same node-path addressing as the rest of the node
-# group (#53/#66). ``from`` is a Python keyword, so the source field is named
-# ``from_node`` and aliased to the wire key ``from``, matching the ``[connection]``
-# key the engine writes; populate-by-name lets the CLI build the model with the
-# Python name while ``--json``/``--schema`` keep the ``from`` wire key.
-_FROM_FIELD = Field(
-    alias="from",
-    description=(
-        "Source node path, relative to the scene root: '.' addresses the root "
-        "itself, 'Player/Arm' a nested node."
-    ),
-)
-_SIGNAL_FIELD = Field(description="The signal name on the source node.")
-_TO_FIELD = Field(
-    description=(
-        "Target node path, relative to the scene root: '.' addresses the root "
-        "itself, 'Player/Arm' a nested node."
-    )
-)
-_METHOD_FIELD = Field(description="The method name on the target node.")
-
-
-class NodeConnectSignalParams(BaseModel):
-    """The operation params of ``gda node connect-signal`` (issue #57).
-
-    Records a connection from a source node's signal to a target node's method,
-    persisted into the ``.tscn`` as a ``[connection]``. As a scene mutation it
-    instantiates the scene (the same trust boundary as ``node set``, ADR-0009).
-    ``path`` is the scene file; ``from``/``signal`` address the source node's
-    signal, ``to``/``method`` the target node's method, by the node group's
-    node-path addressing (#53/#66).
-
-    Contract (issue #57's design decision): the SIGNAL must exist on the source
-    node — a typo or wrong node is a clean ``signal_not_found`` error. The target
-    METHOD need NOT exist: a ``.tscn`` ``[connection]`` is persisted data, and
-    Godot's own editor lets you wire a signal to a not-yet-written method, so the
-    handler can be authored after the wiring — a dangling method is allowed.
-    """
-
-    model_config = ConfigDict(populate_by_name=True, serialize_by_alias=True)
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    from_node: str = _FROM_FIELD
-    signal: str = _SIGNAL_FIELD
-    to: str = _TO_FIELD
-    method: str = _METHOD_FIELD
-
-
-class NodeConnectSignalResult(BaseModel):
-    """The result of ``gda node connect-signal``: the connection it recorded (issue #57).
-
-    Echoes the ``scene_path`` and the four parts of the connection — source node
-    (``from``), ``signal``, target node (``to``), ``method`` — verifiable by
-    reading the saved scene back: the ``[connection]`` now appears in the file.
-    """
-
-    model_config = ConfigDict(populate_by_name=True, serialize_by_alias=True)
-
-    scene_path: str
-    from_node: str = _FROM_FIELD
-    signal: str = _SIGNAL_FIELD
-    to: str = _TO_FIELD
-    method: str = _METHOD_FIELD
-
-
-class NodeDisconnectSignalParams(BaseModel):
-    """The operation params of ``gda node disconnect-signal`` (issue #57).
-
-    Removes an existing signal→method connection from the ``.tscn``. As a scene
-    mutation it instantiates the scene (the same trust boundary as ``node set``,
-    ADR-0009). The four parts address the connection exactly as ``connect-signal``
-    recorded it; a connection that does not exist is a clean ``connection_not_found``
-    error rather than a silent no-op.
-    """
-
-    model_config = ConfigDict(populate_by_name=True, serialize_by_alias=True)
-
-    path: NormalizedPath = Field(description="The .tscn scene file to mutate.")
-    from_node: str = _FROM_FIELD
-    signal: str = _SIGNAL_FIELD
-    to: str = _TO_FIELD
-    method: str = _METHOD_FIELD
-
-
-class NodeDisconnectSignalResult(BaseModel):
-    """The result of ``gda node disconnect-signal``: the connection it removed (issue #57).
-
-    Echoes the ``scene_path`` and the four parts of the removed connection — the
-    same shape as ``connect-signal``'s result — verifiable by reading the saved
-    scene back: the ``[connection]`` is gone from the file.
-    """
-
-    model_config = ConfigDict(populate_by_name=True, serialize_by_alias=True)
-
-    scene_path: str
-    from_node: str = _FROM_FIELD
-    signal: str = _SIGNAL_FIELD
-    to: str = _TO_FIELD
-    method: str = _METHOD_FIELD
 
 
 class ScriptCreateParams(BaseModel):
@@ -2990,10 +2227,11 @@ class ProjectRemoveInputActionResult(BaseModel):
 class GameNode(BaseModel):
     """One node of the RUNNING game's runtime scene tree (Phase 2, ADR-0019).
 
-    The runtime counterpart of :class:`SceneNode`: ``gda game tree`` reports the
-    live ``SceneTree`` after ``_ready`` and dynamic instantiation, so it carries
-    the runtime node ``path`` alongside ``name``/``type``/``children``. Distinct
-    from the on-disk ``.tscn`` read by ``scene get`` (a different object, ADR-0019).
+    The runtime counterpart of :class:`~gda.commands.scene.SceneNode`: ``gda game
+    tree`` reports the live ``SceneTree`` after ``_ready`` and dynamic
+    instantiation, so it carries the runtime node ``path`` alongside
+    ``name``/``type``/``children``. Distinct from the on-disk ``.tscn`` read by
+    ``scene get`` (a different object, ADR-0019).
     """
 
     name: str
